+++ app/src/main/java/com/cajetan/youtubeplus/MainActivity.java
@@ -7,10 +7,13 @@ import android.os.Build;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+import com.cajetan.youtubeplus.utils.YouTubeData;
+
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = MainActivity.class.getSimpleName();
 
     private static final String TEST_VIDEO_ID = "Bcqb7kzekoc";
+
 
     // TODO: implement auto fullscreen on rotation
 
@@ -36,5 +39,11 @@ public class MainActivity extends AppCompatActivity {
         channel.setDescription(getString(R.string.notification_channel_description));
 
         getSystemService(NotificationManager.class).createNotificationChannel(channel);
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
     }
 }