+++ app/src/main/java/com/cajetan/youtubeplus/utils/YouTubeData.java
@@ -1,0 +1,314 @@
+package com.cajetan.youtubeplus.utils;
+
+import android.Manifest;
+import android.accounts.AccountManager;
+import android.app.Activity;
+import android.app.Dialog;
+import android.app.ProgressDialog;
+import android.content.Context;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+import android.os.AsyncTask;
+import android.support.annotation.NonNull;
+import android.text.TextUtils;
+import android.util.Log;
+import android.widget.Toast;
+
+import com.cajetan.youtubeplus.MainActivity;
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.GoogleApiAvailability;
+import com.google.api.client.extensions.android.http.AndroidHttp;
+import com.google.api.client.googleapis.extensions.android.gms.auth.GoogleAccountCredential;
+import com.google.api.client.googleapis.extensions.android.gms.auth.GooglePlayServicesAvailabilityIOException;
+import com.google.api.client.googleapis.extensions.android.gms.auth.UserRecoverableAuthIOException;
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.json.JsonFactory;
+import com.google.api.client.json.jackson2.JacksonFactory;
+import com.google.api.client.util.ExponentialBackOff;
+import com.google.api.services.youtube.YouTube;
+import com.google.api.services.youtube.YouTubeScopes;
+import com.google.api.services.youtube.model.Video;
+import com.google.api.services.youtube.model.VideoListResponse;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+
+import pub.devrel.easypermissions.AfterPermissionGranted;
+import pub.devrel.easypermissions.EasyPermissions;
+
+import static android.app.Activity.RESULT_OK;
+
+public class YouTubeData implements EasyPermissions.PermissionCallbacks {
+
+    GoogleAccountCredential mCredential;
+
+    static final int REQUEST_ACCOUNT_PICKER = 1000;
+    static final int REQUEST_AUTHORIZATION = 1001;
+    static final int REQUEST_GOOGLE_PLAY_SERVICES = 1002;
+    static final int REQUEST_PERMISSION_GET_ACCOUNTS = 1003;
+
+    // TODO: ..
+    private static final String PREF_ACCOUNT_NAME = "accountName";
+    private static final String[] SCOPES = {YouTubeScopes.YOUTUBE_READONLY};
+
+    private String mVideoId = "";
+
+    // TODO ..
+    private Activity mActivity;
+    private Video mResultVideo;
+
+    public YouTubeData(Activity parentActivity) {
+        mCredential = GoogleAccountCredential.usingOAuth2(
+                parentActivity.getApplicationContext(), Arrays.asList(SCOPES))
+                .setBackOff(new ExponentialBackOff());
+
+        mActivity = parentActivity;
+    }
+
+    public Video getVideoDataById(String videoId) {
+        mVideoId = videoId;
+        getResultsFromApi();
+
+        return mResultVideo;
+    }
+
+    // TODO: tons all refactoring all over this class
+    private void getResultsFromApi() {
+        if (!isGooglePlayServicesAvailable()) {
+            acquireGooglePlayServices();
+        } else if (mCredential.getSelectedAccountName() == null) {
+            chooseAccount();
+        } else if (!isDeviceOnline()) {
+            mResultVideo = null;
+        } else {
+            new VideoDataTask(mCredential).execute(mVideoId);
+        }
+    }
+
+    /**
+     * An asynchronous task that handles the YouTube Data API call.
+     */
+    private class VideoDataTask extends AsyncTask<String, Void, Video> {
+        private com.google.api.services.youtube.YouTube mService = null;
+        private Exception mLastError = null;
+
+        VideoDataTask(GoogleAccountCredential credential) {
+            HttpTransport transport = AndroidHttp.newCompatibleTransport();
+            JsonFactory jsonFactory = JacksonFactory.getDefaultInstance();
+            mService = new com.google.api.services.youtube.YouTube.Builder(
+                    transport, jsonFactory, credential)
+                    .setApplicationName("YouTube Plus")
+                    .build();
+        }
+
+        /**
+         * Background task to call YouTube Data API.
+         *
+         * @param params no parameters needed for this task.
+         */
+        @Override
+        protected Video doInBackground(String... params) {
+            try {
+                return getDataFromApi(params[0]);
+            } catch (Exception e) {
+                Log.e("MainActivityException", "Exception: " + e.toString());
+                mLastError = e;
+                cancel(true);
+                return null;
+            }
+        }
+
+        /**
+         * Fetch information about the requested video.
+         *
+         * @return List of Strings containing information about the channel.
+         * @throws IOException
+         */
+        private Video getDataFromApi(String videoId) throws IOException {
+
+            HashMap<String, String> parameters = new HashMap<>();
+            parameters.put("part", "snippet,contentDetails,statistics");
+            parameters.put("id", "Bcqb7kzekoc");
+
+            YouTube.Videos.List videosListByIdRequest = mService.videos().list(parameters.get("part"));
+            if (parameters.containsKey("id") && !parameters.get("id").equals(""))
+                videosListByIdRequest.setId(parameters.get("id"));
+
+            VideoListResponse response = videosListByIdRequest.execute();
+            return response.getItems().get(0);
+        }
+
+        @Override
+        protected void onPreExecute() {
+            mResultVideo = null;
+        }
+
+        @Override
+        protected void onPostExecute(Video output) {
+            mResultVideo = output;
+        }
+
+        @Override
+        protected void onCancelled() {
+            if (mLastError != null) {
+                if (mLastError instanceof GooglePlayServicesAvailabilityIOException) {
+                    // TODO: fix too
+//                    showGooglePlayServicesAvailabilityErrorDialog(
+//                            ((GooglePlayServicesAvailabilityIOException) mLastError)
+//                                    .getConnectionStatusCode());
+                } else if (mLastError instanceof UserRecoverableAuthIOException) {
+                    mActivity.startActivityForResult(
+                            ((UserRecoverableAuthIOException) mLastError).getIntent(),
+                            YouTubeData.REQUEST_AUTHORIZATION);
+                } else {
+                    Log.e("YouTubeData", "Error ocurred: " + mLastError.getMessage());
+                    mResultVideo = null;
+                }
+            } else {
+                mResultVideo = null;
+                Log.e("YouTubeData", "Request cancelled");
+            }
+        }
+    }
+
+    public void onParentActivityResult(int requestCode, int resultCode, Intent data) {
+        switch (requestCode) {
+            case REQUEST_GOOGLE_PLAY_SERVICES:
+                if (resultCode != RESULT_OK) {
+                    Toast.makeText(mActivity,
+                            "This app requires Google Play Services. Please install " +
+                                    "Google Play Services on your device and relaunch this app.",
+                            Toast.LENGTH_LONG).show();
+                } else {
+                    getResultsFromApi();
+                }
+                break;
+            case REQUEST_ACCOUNT_PICKER:
+                if (resultCode == RESULT_OK && data != null &&
+                        data.getExtras() != null) {
+                    String accountName =
+                            data.getStringExtra(AccountManager.KEY_ACCOUNT_NAME);
+                    if (accountName != null) {
+                        SharedPreferences settings =
+                                mActivity.getPreferences(Context.MODE_PRIVATE);
+                        SharedPreferences.Editor editor = settings.edit();
+                        editor.putString(PREF_ACCOUNT_NAME, accountName);
+                        editor.apply();
+                        mCredential.setSelectedAccountName(accountName);
+                        getResultsFromApi();
+                    }
+                }
+                break;
+            case REQUEST_AUTHORIZATION:
+                if (resultCode == RESULT_OK) {
+                    getResultsFromApi();
+                }
+                break;
+        }
+    }
+
+    @AfterPermissionGranted(REQUEST_PERMISSION_GET_ACCOUNTS)
+    private void chooseAccount() {
+        if (EasyPermissions.hasPermissions(
+                mActivity, Manifest.permission.GET_ACCOUNTS)) {
+            String accountName = mActivity.getPreferences(Context.MODE_PRIVATE)
+                    .getString(PREF_ACCOUNT_NAME, null);
+            if (accountName != null) {
+                mCredential.setSelectedAccountName(accountName);
+                getResultsFromApi();
+            } else {
+                // Start a dialog from which the user can choose an account
+                mActivity.startActivityForResult(
+                        mCredential.newChooseAccountIntent(),
+                        REQUEST_ACCOUNT_PICKER);
+            }
+        } else {
+            // Request the GET_ACCOUNTS permission via a user dialog
+            EasyPermissions.requestPermissions(
+                    mActivity,
+                    "This app needs to access your Google account (via Contacts).",
+                    REQUEST_PERMISSION_GET_ACCOUNTS,
+                    Manifest.permission.GET_ACCOUNTS);
+        }
+    }
+
+    /**
+     * Checks whether the device currently has a network connection.
+     *
+     * @return true if the device has a network connection, false otherwise.
+     */
+    private boolean isDeviceOnline() {
+        ConnectivityManager connMgr =
+                (ConnectivityManager) mActivity.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo networkInfo = connMgr.getActiveNetworkInfo();
+        return (networkInfo != null && networkInfo.isConnected());
+    }
+
+    /**
+     * Check that Google Play services APK is installed and up to date.
+     *
+     * @return true if Google Play Services is available and up to
+     * date on this device; false otherwise.
+     */
+    private boolean isGooglePlayServicesAvailable() {
+        GoogleApiAvailability apiAvailability =
+                GoogleApiAvailability.getInstance();
+        final int connectionStatusCode =
+                apiAvailability.isGooglePlayServicesAvailable(mActivity);
+        return connectionStatusCode == ConnectionResult.SUCCESS;
+    }
+
+    /**
+     * Attempt to resolve a missing, out-of-date, invalid or disabled Google
+     * Play Services installation via a user dialog, if possible.
+     */
+    private void acquireGooglePlayServices() {
+        GoogleApiAvailability apiAvailability =
+                GoogleApiAvailability.getInstance();
+        final int connectionStatusCode =
+                apiAvailability.isGooglePlayServicesAvailable(mActivity);
+        if (apiAvailability.isUserResolvableError(connectionStatusCode)) {
+            // TODO: fix it
+//            showGooglePlayServicesAvailabilityErrorDialog(connectionStatusCode);
+        }
+    }
+
+
+//    /**
+//     * Display an error dialog showing that Google Play Services is missing
+//     * or out of date.
+//     *
+//     * @param connectionStatusCode code describing the presence (or lack of)
+//     *                             Google Play Services on this device.
+//     */
+//    void showGooglePlayServicesAvailabilityErrorDialog(
+//            final int connectionStatusCode) {
+//        GoogleApiAvailability apiAvailability = GoogleApiAvailability.getInstance();
+//        Dialog dialog = apiAvailability.getErrorDialog(
+//                MainActivity.this,
+//                connectionStatusCode,
+//                REQUEST_GOOGLE_PLAY_SERVICES);
+//        dialog.show();
+//    }
+
+    @Override
+    public void onPermissionsGranted(int requestCode, List<String> perms) {
+
+    }
+
+    @Override
+    public void onPermissionsDenied(int requestCode, List<String> perms) {
+
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        EasyPermissions.onRequestPermissionsResult(
+                requestCode, permissions, grantResults, this);
+    }
+}